+++ example/demo/concent-layout/node.js
@@ -10,7 +10,6 @@ class BaseNode extends Node {
     this.options = opts;
   }
   draw = (opts) => {
-    console.log(opts)
     let container = $('<div class="decision-new-node"></div>')
                     .attr('id', opts.id)
                     .css('top', opts.top + 'px')
