+++ example/demo/fruchterman/index.less
@@ -1,0 +1,11 @@
+.fruchterman-tree-page {
+    position: relative;
+    height: 100%;
+    width: 100%;
+    .fruchterman-tree-canvas{
+      height: 100%;
+      width: 100%;
+      background: #fff;
+    }
+    
+}
