+++ example/demo/fruchterman/node.less
@@ -1,0 +1,10 @@
+.fruchterman-node {
+    position: absolute;
+    width: 30px;
+    height: 30px;
+    line-height: 30px;
+    text-align: center;
+    border-radius: 50%;
+    border: 1px solid #999;
+    background: #C6E5Ff;
+  }
