+++ src/canvas/baseCanvas.js
@@ -3163,7 +3163,6 @@ class BaseCanvas extends Canvas {
           }
         });
       } else if(_.get(this.layout, 'type') === 'gridLayout') {
-        
         const _opts = $.extend({
           // 布局画布总宽度
           width: 150,
@@ -3191,11 +3190,53 @@ class BaseCanvas extends Canvas {
             strength: 1
           }
         }, _.get(this.layout, 'options'), true);
-
         // 自动布局
         if (_.get(this.layout, 'type') === 'gridLayout') {
 
           Layout.gridLayout({
+            opts: _opts,
+            data: {
+              groups: data.groups,
+              nodes: data.nodes,
+              // 加工线条数据，兼容endpoint为id的属性，d3没这个概念
+              edges: data.edges.map(item => ({
+                source: item.type === 'endpoint' ? item.sourceNode : item.source,
+                target: item.type === 'endpoint' ? item.targetNode : item.target
+              }))
+            }
+          })
+        }
+      } else if(_.get(this.layout, 'type') === 'fruchterman') {
+        const _opts = $.extend({
+           // 布局画布总宽度
+           width,
+           // 布局画布总长度
+           height,
+            /** 停止迭代的最大迭代数 */
+          maxIteration: 1000,
+            /** 布局中心 */
+          center: [width / 2, height / 2],
+            /** 重力大小，影响图的紧凑程度 */
+          gravity: 5,
+          /** 速度 */
+          speed: 5,
+           /** 是否产生聚类力 */
+          clustering: false,
+           /** 聚类力大小 */
+          clusterGravity: 10,
+          link: {
+            // 以node的什么字段为寻找id，跟d3原理一样
+            id: 'id',
+            // 线条的距离
+            distance: 100,
+            // 线条的粗细
+            strength: 1
+          }
+        }, _.get(this.layout, 'options'), true);
+        // 自动布局
+        if (_.get(this.layout, 'type') === 'fruchterman') {
+
+          Layout.fruchterman({
             opts: _opts,
             data: {
               groups: data.groups,
